fortune_wheel/requirements.txt
aiogrpc==1.8
alabaster==0.7.12
-e git+https://github.com/Lalluviadel/Vector-robot.git@975dd7eb3f129771b7288695f5640ed332fa4fee#egg=anki_vector
astroid==2.13.2
Babel==2.11.0
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
cryptography==39.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docutils==0.19
Flask==2.2.2
Flask-Babel==0.8
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.4
flup==1.0.3
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.51.1
humanfriendly==10.0
idna==3.4
imagesize==1.4.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.24.1
packaging==23.0
pbr==5.11.1
Pillow==9.4.0
platformdirs==2.6.2
protobuf==3.20.1
pycparser==2.21
Pygments==2.14.0
pylint==2.15.10
PyOpenGL==3.1.6
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
scipy==1.10.0
six==1.16.0
snowballstemmer==2.2.0
speaklater==1.3
Sphinx==6.1.3
sphinx-rtd-theme==0.5.1
sphinx_autodoc_typehints==1.20.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib.applehelp==1.0.3
SQLAlchemy==1.4.46
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
Tempita==0.5.2
termcolor==2.2.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
